+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to the "zephyr-workbench" extension will be documented in th
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [2.6.5]
+
+### Install Runners
+- Added OpenOCD Infineon installation support on Linux and Windows
+
 ## [2.6.4]
 
 ### Snippets